requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0
tqdm==4.66.4
requests==2.32.3
sentence-transformers==3.0.1
transformers==4.42.4
torch>=2.1.0
faiss-cpu==1.8.0
scikit-learn==1.4.2
pytest==8.3.1
httpx==0.27.0
